fix KeyError bug

+++ base/music_base.py
@@ -84,14 +84,13 @@ class MusicPlayer(CallBase):
         playlist = self.playlist
         chat_id = cb.message.chat.id
         if playlist:
-            if len(playlist[chat_id]) >= 1:
+            if chat_id in playlist and len(playlist[chat_id]) >= 1:
                 self.extend_playlist(
                     user_id, chat_id, title, duration, yt_url, yt_id, "music"
                 )
                 mess = await cb.edit_message_text(gm(chat_id, "track_queued"))
                 await sleep(5)
                 return await mess.delete()
-            pass
         messy = await cb.edit_message_text(gm(chat_id, "process"))
         audio_url = get_audio_direct_link(yt_url)
         try:
